soa-football-app/requirements.txt
fastapi[all]
uvicorn
dependency_injector
containers
pydantic
python-dotenv
requests
